move ValidateJoinRequest
